# Map.Domain/Models/Auth/LoginDto.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace Map.Domain.Models.Auth;
public class LoginDto
{
    public string Email { get; set; }
    public string Password { get; set; }
}